+++ core/initializer/parallelism/test_executor.go
@@ -8,11 +8,13 @@ import (
 	"github.com/docker/go-connections/nat"
 	"github.com/kurtosis-tech/kurtosis/commons/docker"
 	"github.com/kurtosis-tech/kurtosis/commons/networks"
+	"github.com/kurtosis-tech/kurtosis/commons/testsuite"
 	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"
 	"io"
 	"io/ioutil"
 	"os"
+	"time"
 )
 
 /*
@@ -40,24 +42,126 @@ const (
 	testImageNameArg        = "TEST_IMAGE_NAME"
 	testControllerIpArg     = "TEST_CONTROLLER_IP"
 	testVolumeMountpointArg = "TEST_VOLUME_MOUNTPOINT"
+
+	// After we hard-timeout a test, how long we'll give the test to clean itself up (namely the Docker network & containers)
+	//  before we call it lost and continue on
+	networkTeardownGraceTime = 60 * time.Second
+
+	// When we're tearing down a network after a test (either after normal exit or test timeout), this is the maximum
+	//  time we'll wait for each container to stop
+	networkTeardownContainerStopTimeout = 10 * time.Second
 )
 
-type testExecutor struct {
-	log *logrus.Logger
+// Because a test is run in its own goroutine to allow us to time it out, we need to pass the results back
+//  via a channel. Tihs struct is what's passed over the chanel.
+type testResult struct {
+	// Whether the test passed or not (undefined if an error occurred that prevented us from retrieving test results)
+	testPassed   bool
+
+	// If not nil, the error that prevented us from retrieving the test result
+	executionErr error
 }
 
-func newTestExecutor(log *logrus.Logger) *testExecutor {
-	return &testExecutor{log: log}
+// Executor responsible for running a test, with timeout, cleaning up after the test as needed
+type testExecutor struct {
+	log *logrus.Logger
+
+	// Tests already declare timeouts, but that timeout only represents time spent running the actual test
+	// This value represents how much time *on top of* the test timeout we'll give to the test goroutine for it to do
+	//  do test setup and finalization (and if this total timeout is exceeded, we'll try to shut down everything in
+	//  the test - the controller, the containers in the network, the Docker network itself, etc.)
+	additionalTestTimeoutBuffer time.Duration
+}
+
+func newTestExecutor(log *logrus.Logger, additionalTestTimeoutBuffer time.Duration) *testExecutor {
+	return &testExecutor{
+		log: log,
+		additionalTestTimeoutBuffer: additionalTestTimeoutBuffer,
+	}
+}
+
+// TODO Just make all these params struct params - passing them through each to function is tedious and error-prone because
+//  we have a lot of string params and it's easy to get them mixed up
+/*
+Returns:
+	bool: A boolean indicating if the test passed (will be undefined if the test result couldn't be retrieved for any reason)
+	error: If not nil, represents the error hit while running the test that prevented the retrieval of the test result
+ */
+func (executor testExecutor) runTest(
+		executionInstanceId uuid.UUID,
+		dockerClient *client.Client,
+		subnetMask string,
+		testControllerImageName string,
+		testControllerLogLevel string,
+		testServiceImageName string,
+		testControllerEnvVars map[string]string,
+		testName string,
+		test testsuite.Test) (bool, error) {
+	testResultChan := make(chan testResult)
+
+	// When this is breached, we'll try to tear down everything
+	totalTimeout := test.GetTimeout() + executor.additionalTestTimeoutBuffer
+
+	context, cancelFunc := context.WithCancel(context.Background())
+	defer cancelFunc()
+
+	// We run the test in a separate goroutine because we don't know if the test will even respect the context we pass in -
+	//  we hope so, but (because this runs user-written code) we can't trust it so we give ourselves the option to move
+	//  on if the test, e.g., infinite-loops
+	go func() {
+		testPassed, setupErr := executor.runTestGoroutine(
+			context,
+			executionInstanceId,
+			dockerClient,
+			subnetMask,
+			testControllerImageName,
+			testControllerLogLevel,
+			testServiceImageName,
+			testControllerEnvVars,
+			testName)
+		testResultChan <- testResult{
+			testPassed:   testPassed,
+			executionErr: setupErr,
+		}
+	}()
+
+	var timedOut bool
+	var testExecutionResult testResult
+	select {
+	case testExecutionResult = <- testResultChan:
+		timedOut = false
+	case <- time.After(totalTimeout):
+		timedOut = true
+	}
+
+	if timedOut {
+		executor.log.Tracef("Hit hard test timeout of %v; the context is being cancelled to give it the chance to exit gracefully...", totalTimeout)
+		cancelFunc()
+
+		// We've now cancelled the context so the test goroutine *should* exit gracefully soon
+		select {
+		case testExecutionResult = <- testResultChan:
+			executor.log.Info("Test goroutine exited gracefully after context cancellation")
+		case <- time.After(networkTeardownGraceTime):
+			executor.log.Warnf(
+				"Test goroutine didn't exit gracefully after context cancellation even after a grace period of %v; the test goroutine is being called lost",
+				networkTeardownGraceTime,
+			)
+		}
+		return false, stacktrace.NewError("Test hit hard timeout, %v", totalTimeout)
+	} else {
+		return testExecutionResult.testPassed, testExecutionResult.executionErr
+	}
 }
 
 /*
 Returns:
-	error: An error if an error occurred *while setting up or running the test* (independent from whether the test itself passed)
+	error: If an error occurred that prevented us from running the test & retrieving the results (independent from whether the test itself passed)
 	bool: A boolean indicating whether the test passed (undefined if an error occurred running the test)
- */
-func (executor testExecutor) runTest(
+*/
+func (executor testExecutor) runTestGoroutine(
+		context context.Context,
 		executionInstanceId uuid.UUID,
-		testContext context.Context,
 		dockerClient *client.Client,
 		subnetMask string,
 		testControllerImageName string,
@@ -66,7 +170,9 @@ func (executor testExecutor) runTest(
 		testControllerEnvVars map[string]string,
 		testName string) (bool, error) {
 	executor.log.Info("Creating Docker manager from environment settings...")
-	dockerManager, err := docker.NewDockerManager(executor.log, testContext, dockerClient)
+	// NOTE: at this point, all Docker commands from here forward will be bound by the Context that we pass in here - we'll
+	//  only need to cancel this context once
+	dockerManager, err := docker.NewDockerManager(executor.log, dockerClient)
 	if err != nil {
 		return false, stacktrace.Propagate(err, "An error occurred getting the Docker manager for test %v", testName)
 	}
@@ -82,7 +188,7 @@ func (executor testExecutor) runTest(
 	if err != nil {
 		return false, stacktrace.Propagate(err, "An error occurred getting the gateway IP")
 	}
-	_, err = dockerManager.CreateNetwork(networkName, subnetMask, gatewayIp)
+	_, err = dockerManager.CreateNetwork(context, networkName, subnetMask, gatewayIp)
 	if err != nil {
 		return false, stacktrace.Propagate(err, "Error occurred creating Docker network %v for test %v", networkName, testName)
 	}
@@ -95,6 +201,7 @@ func (executor testExecutor) runTest(
 		return false, stacktrace.NewError("An error occurred getting an IP for the test controller")
 	}
 	testPassed, err := runControllerContainer(
+		context,
 		executor.log,
 		dockerManager,
 		networkName,
@@ -115,6 +222,7 @@ func (executor testExecutor) runTest(
 	return testPassed, nil
 }
 
+// TODO Move many of these args to the struct itself to simplify the params of this function
 /*
 Helper function to run the controller container against the given test network.
 
@@ -138,6 +246,7 @@ Returns:
 	error: if any error occurred during the execution of the controller (independent of the test itself)
 */
 func runControllerContainer(
+			context context.Context,
 			log *logrus.Logger,
 			manager *docker.DockerManager,
 			networkName string,
@@ -152,7 +261,7 @@ func runControllerContainer(
 			executionUuid uuid.UUID) (bool, error){
 	volumeName := fmt.Sprintf("%v-%v", executionUuid.String(), testName)
 	log.Debugf("Creating Docker volume %v which will be shared with the test network...", volumeName)
-	if err := manager.CreateVolume(volumeName); err != nil {
+	if err := manager.CreateVolume(context, volumeName); err != nil {
 		return false, stacktrace.Propagate(err, "Error creating Docker volume to share amongst test nodes")
 	}
 	log.Debugf("Docker volume %v created successfully", volumeName)
@@ -181,29 +290,33 @@ func runControllerContainer(
 	}
 	log.Debugf("Environment variables that are being passed to the controller: %v", envVariables)
 
+	bindMounts := map[string]string{
+		// Because the test controller will need to spin up new images, we need to bind-mount the host Docker engine into the test controller
+		"/var/run/docker.sock": "/var/run/docker.sock",
+		logTmpFile.Name():      controllerLogMountFilepath,
+	}
+
+	volumeMounts := map[string]string{
+		volumeName: testVolumeMountpoint,
+	}
+
 	_, controllerContainerId, err := manager.CreateAndStartContainer(
+		context,
 		controllerImageName,
 		networkName,
 		controllerIpAddr,
 		make(map[nat.Port]bool),
 		nil, // The controller image's CMD should be parameterized, so we don't specify a start command here
 		envVariables,
-		map[string]string{
-			// Because the test controller will need to spin up new images, we need to bind-mount the host Docker engine into the test controller
-			"/var/run/docker.sock": "/var/run/docker.sock",
-			logTmpFile.Name():      controllerLogMountFilepath,
-		},
-		map[string]string{
-			volumeName: testVolumeMountpoint,
-		})
+		bindMounts,
+		volumeMounts)
 	if err != nil {
 		return false, stacktrace.Propagate(err, "Failed to run test controller container")
 	}
 	log.Infof("Controller container started successfully with id %s", controllerContainerId)
 
 	log.Info("Waiting for controller container to exit...")
-	// TODO add a timeout here if the test doesn't complete successfully
-	exitCode, err := manager.WaitForExit(controllerContainerId)
+	exitCode, err := manager.WaitForExit(context, controllerContainerId)
 	if err != nil {
 		return false, stacktrace.Propagate(err, "Failed when waiting for controller to exit")
 	}
@@ -229,10 +342,12 @@ as a deferred function.
 */
 func removeNetworkDeferredFunc(log *logrus.Logger, dockerManager *docker.DockerManager, networkName string) {
 	log.Infof("Attempting to remove Docker network with name %v...", networkName)
-	err := dockerManager.RemoveNetwork(networkName)
-	if err != nil {
+	// We use the background context here because we want to try and tear down the network even if the context the test was running in
+	//  was cancelled. This might not be right - the right way to do it might be to pipe a separate context for the network teardown to here!
+	if err := dockerManager.RemoveNetwork(context.Background(), networkName, networkTeardownContainerStopTimeout); err != nil {
 		log.Errorf("An error occurred removing Docker network with name %v:", networkName)
 		log.Error(err.Error())
+		log.Error("NOTE: This means you will need to clean up the Docker network manually!!")
 	} else {
 		log.Infof("Docker network %v successfully removed", networkName)
 	}