+++ core/initializer/test_suite_runner.go
@@ -15,6 +15,7 @@ import (
 	"net"
 	"os"
 	"sort"
+	"time"
 )
 
 // =============================== "enum" for test result =========================================
@@ -34,6 +35,9 @@ type TestSuiteRunner struct {
 
 	// The test controller image-specific string representing the log level, that will be passed as-is to the test controller
 	testControllerLogLevel	string
+
+	// The additional time, on top of the declared per-test timeout, that's given to tests for setup & teardown
+	additionalTestTimeoutBuffer time.Duration
 }
 
 const (
@@ -56,13 +60,17 @@ func NewTestSuiteRunner(
 			testServiceImageName string,
 			testControllerImageName string,
 			testControllerLogLevel string,
-			testControllerEnvVars map[string]string) *TestSuiteRunner {
+			testControllerEnvVars map[string]string,
+			// TODO Move this extra setup/teardown timeout buffer to be something test-specific (since it will depend on
+			//  the network the test is spinning up)
+			additionalTestTimeoutBuffer time.Duration) *TestSuiteRunner {
 	return &TestSuiteRunner{
 		testSuite:               testSuite,
 		testServiceImageName:    testServiceImageName,
 		testControllerImageName: testControllerImageName,
 		testControllerLogLevel:  testControllerLogLevel,
 		testControllerEnvVars:   testControllerEnvVars,
+		additionalTestTimeoutBuffer: additionalTestTimeoutBuffer,
 	}
 }
 
@@ -82,12 +90,13 @@ func (runner TestSuiteRunner) RunTests(testNamesToRun []string, testParallelism 
 	}
 
 	// Validate all the requested tests exist
-	testsToRun := make(map[string]bool)
+	testsToRun := make(map[string]testsuite.Test)
 	for _, testName := range testNamesToRun {
-		if _, found := allTests[testName]; !found {
+		test, found := allTests[testName]
+		if !found {
 			return false, stacktrace.NewError("No test registered with name '%v'", testName)
 		}
-		testsToRun[testName] = true
+		testsToRun[testName] = test
 	}
 
 	executionInstanceId := uuid.Generate()
@@ -109,13 +118,20 @@ func (runner TestSuiteRunner) RunTests(testNamesToRun []string, testParallelism 
 		runner.testControllerLogLevel,
 		runner.testServiceImageName,
 		runner.testControllerEnvVars,
-		testParallelism)
+		testParallelism,
+		runner.additionalTestTimeoutBuffer)
 
 	logrus.Infof("Running %v tests with execution ID %v...", len(testsToRun), executionInstanceId.String())
-	testOutputs := testExecutor.RunInParallel(testParams)
+	interceptor := parallelism.NewErroneousSystemLogCaptureWriter()
+	testOutputs := testExecutor.RunInParallel(interceptor, testParams)
 
 	logrus.Infof("Printing results for %v tests...", len(testsToRun))
 	allTestsPassed = processTestOutputs(testsToRun, testOutputs)
+
+	// If there was any erroneous system-level logging during parallel test execution, loudly display that to the user
+	capturedErroneousMessages := interceptor.GetCapturedMessages()
+	logErroneousSystemLogging(capturedErroneousMessages)
+
 	return allTestsPassed, nil
 }
 
@@ -125,7 +141,7 @@ Helper function to build, from the set of tests to run, the map of test params t
 Args:
 	testsToRun: A "set" of test names to run in parallel
  */
-func buildTestParams(executionInstanceId uuid.UUID, testsToRun map[string]bool) (map[string]parallelism.ParallelTestParams, error) {
+func buildTestParams(executionInstanceId uuid.UUID, testsToRun map[string]testsuite.Test) (map[string]parallelism.ParallelTestParams, error) {
 
 	subnetMaskBits := BITS_IN_IP4_ADDR - NETWORK_WIDTH_BITS
 
@@ -146,7 +162,7 @@ func buildTestParams(executionInstanceId uuid.UUID, testsToRun map[string]bool) 
 
 	testIndex := 0
 	testParams := make(map[string]parallelism.ParallelTestParams)
-	for testName, _ := range testsToRun {
+	for testName, test := range testsToRun {
 		// Pick the next free available subnet IP, considering all the tests we've started previously
 		subnetIpInt := subnetStartIpInt + uint32(testIndex) * uint32(math.Pow(2, NETWORK_WIDTH_BITS))
 		subnetIp := make(net.IP, 4)
@@ -160,12 +176,7 @@ func buildTestParams(executionInstanceId uuid.UUID, testsToRun map[string]bool) 
 		}
 		logrus.Tracef("Temp file to write logs for test %v to: %v", testName, tempFp.Name())
 
-		testParams[testName] = parallelism.ParallelTestParams{
-			TestName:            testName,
-			LogFp:               tempFp,
-			SubnetMask:          subnetCidrStr,
-			ExecutionInstanceId: executionInstanceId,
-		}
+		testParams[testName] = *parallelism.NewParallelTestParams(testName, test, tempFp, subnetCidrStr, executionInstanceId)
 		testIndex++
 	}
 	return testParams, nil
@@ -181,7 +192,7 @@ Args:
 Returns:
 	True if all tests passed, false otherwise
  */
-func processTestOutputs(testsToRun map[string]bool, testOutputs map[string]parallelism.ParallelTestOutput) bool {
+func processTestOutputs(testsToRun map[string]testsuite.Test, testOutputs map[string]parallelism.ParallelTestOutput) bool {
 	// We want normalized output between runs of the tests suite so we sort the tests by name
 	testPrintOrder := []string{}
 	for testName, _ := range testsToRun {
@@ -191,7 +202,7 @@ func processTestOutputs(testsToRun map[string]bool, testOutputs map[string]paral
 
 	allTestResults := make(map[string]testResult)
 	for _, name := range testPrintOrder {
-		logrus.Infof("---------------------------------- %v --------------------------------", name)
+		printBanner(name, false)
 
 		output := testOutputs[name]
 		passed := output.TestPassed
@@ -219,7 +230,7 @@ func processTestOutputs(testsToRun map[string]bool, testOutputs map[string]paral
 		allTestResults[name] = result
 	}
 
-	logrus.Info("================================== TEST RESULTS ================================")
+	printBanner("All Test Results", false)
 	allTestsPassed := true
 	for _, testName := range testPrintOrder {
 		result := allTestResults[testName]
@@ -258,3 +269,50 @@ func logTestResult(testName string, executionErr error, testPassed bool) testRes
 	return result
 }
 
+func printBanner(contents string, isError bool) {
+	bannerString := "=================================================================================================="
+	contentString := fmt.Sprintf("                              %v", contents)
+	if !isError {
+		logrus.Info("")
+		logrus.Info(bannerString)
+		logrus.Info(contentString)
+		logrus.Info(bannerString)
+	} else {
+		logrus.Error("")
+		logrus.Error(bannerString)
+		logrus.Error(contentString)
+		logrus.Error(bannerString)
+	}
+}
+
+/*
+Helper function to print a big warning if there was logging to the system-level logging when there should only have been
+ logging to the test-specific logger
+ */
+func logErroneousSystemLogging(capturedErroneousMessages []parallelism.ErroneousSystemLogInfo) {
+	if len(capturedErroneousMessages) == 0 {
+		return
+	}
+
+	printBanner("Erroneous Logs", true)
+	logrus.Error("There were log messages printed to the system-level logger during parallel test execution!")
+	logrus.Error("Because the system-level logger is shared and the tests run in parallel, the messages cannot be")
+	logrus.Error(" attributed to any specific test. This is:")
+	logrus.Error("   1) A bug in Kurtosis, and a system-level logger call was used when a test-specific logger")
+	logrus.Error("       should have been used (likely)")
+	logrus.Error("   2) Third-party code calling logrus independently, and there's nothing we can do (unlikely, but possible)")
+	logrus.Error("")
+	logrus.Error("The log message(s) attempted, and the stacktrace(s) of origination, are as follows in the order they were logged:")
+	logrus.Error("")
+
+	for i, messageInfo := range capturedErroneousMessages {
+		logrus.Errorf("----------------- Erroneous Message #%d -------------------", i+1)
+		logrus.Error("Message:")
+		logrus.StandardLogger().Out.Write(messageInfo.Message)
+		logrus.StandardLogger().Out.Write([]byte("\n")) // The message likely won't come with a newline so we add it
+		logrus.Error("")
+		logrus.Error("Stacktrace:")
+		logrus.StandardLogger().Out.Write(messageInfo.Stacktrace)
+		logrus.StandardLogger().Out.Write([]byte("\n")) // The stacktrace likely won't end with a newline so we add it
+	}
+}